reorganise components
https://coveord.atlassian.net/browse/KIT-1622
